requirements.txt
djangorestframework
django-filter
markdown